Ensured correct display of performance marker times
Float granularity is too low for performance markers - too many markers
were showing the same time. Upgraded it to a double - this allows the
multiplication to generate the right markers in the network client

+++ adaptors/base/performance-logging/networking/network-performance-client.cpp
@@ -170,12 +170,13 @@ bool NetworkPerformanceClient::TransmitMarker( const PerformanceMarker& marker, 
     // write out the time stamp
     char buffer[64];
     int size = snprintf( buffer, sizeof(buffer),"%.6f (seconds), %s\n",
-                                    (float)( marker.GetTimeStamp().microseconds * MICROSECONDS_TO_SECOND ),
-                                    description );
+                         ( (double) marker.GetTimeStamp().microseconds * MICROSECONDS_TO_SECOND ),
+                         description );
 
    return mSocket->Write( buffer, size );
 
   }
+
 
   // todo serialize the data
   return false;
